Adding reformatting hints.
When the format string includes a format, then use it as a hint
for the initial string representation.  This is a lot of parsing
since it goes from value -> string -> value -> formatted string.

This can also lead to some issues, like if you want to print an
integer as a hex number, but the integer also happens to be a valid
bit_vector.  In this case, the output might be incorrect.