Updating package names, splitting packages, renaming fpr.
Updating the package names to add _p to the end of them.  This
allows the package to be named fmt_p which I personally like.

Splitting string_list type to it's own package called string_list_p.
This is used for convenience sake by the user if they want, but
isn't required to be included in fmt_p.

Adding colors_p which defines the set of ANSI color codes.
This is just a convenience package which may be utilized in the future.

Renaming fpr to fmt.  It seemed natural to name the function that
actually formats the string to fmt which is what some other languages
name.